# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk konsumsi TMDB API, manajemen data film, dashboard analytics, dan chart visualization",
        "title": "Movie Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/yourusername/movie-backend",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8010",
    "basePath": "/api/v1",
    "paths": {
        "/charts": {
            "get": {
                "description": "Get combined pie chart (by language) and column chart (by year) data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get chart data for visualization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/charts/column": {
            "get": {
                "description": "Get movie distribution by release year for column chart visualization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get column chart data by year",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Column chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve column chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/charts/monthly/{year}": {
            "get": {
                "description": "Get movie distribution by month for a specific year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get monthly chart data for a specific year",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Year (e.g., 2024)",
                        "name": "year",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid year",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve monthly chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/charts/pie": {
            "get": {
                "description": "Get movie distribution by original language for pie chart visualization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charts"
                ],
                "summary": "Get pie chart data by language",
                "responses": {
                    "200": {
                        "description": "Pie chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve pie chart data",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/stats": {
            "get": {
                "description": "Get comprehensive dashboard analytics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get dashboard statistics",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve statistics",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/movies": {
            "get": {
                "description": "Get list of all movies with pagination, search, sorting, and date range filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get all movies",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title or overview",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "updated_at",
                        "description": "Sort by field (id, title, release_date, vote_average, popularity, created_at, updated_at)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "DESC",
                        "description": "Sort order (ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of movies",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new movie entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Create a new movie",
                "parameters": [
                    {
                        "description": "Movie object",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Movie created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Get a single movie by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get movie by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie details",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid movie ID",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Update a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Movie object",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Movie"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a movie by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Delete a movie",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid movie ID",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/sync/last-log": {
            "get": {
                "description": "Get the most recent sync operation log",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Get last sync log",
                "responses": {
                    "200": {
                        "description": "Last sync log",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve sync log",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        },
        "/sync/movies": {
            "post": {
                "description": "Fetch and sync popular movies from TMDB API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "Sync movies from TMDB",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Number of pages to sync (1-10)",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sync completed successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Sync failed",
                        "schema": {
                            "$ref": "#/definitions/utils.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "utils.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}